fix: double-mangled paths and clippy cast warnings
- Fix rustledger_core::rustledger_core:: double paths in extract_cmd
- Replace 'as i32'/'as i64' with i32::from()/i64::from() for clippy

Verified locally: fmt, clippy, test all pass.

+++ crates/rustledger-query/src/executor/mod.rs
@@ -2319,8 +2319,8 @@ impl<'a> Executor<'a> {
             };
 
             let year = Value::Integer(i64::from(txn.date.year()));
-            let month = Value::Integer(i64::from(txn.date.month() as i32));
-            let day = Value::Integer(i64::from(txn.date.day() as i32));
+            let month = Value::Integer(i64::from(txn.date.month()));
+            let day = Value::Integer(i64::from(txn.date.day()));
 
             for posting in &txn.postings {
                 // Update running balance
